+++ lib/theme/vit_combo_box_theme_data.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/widgets.dart';
+
+class VitComboBoxThemeData {
+  final RenderBox? Function()? parentRenderBoxGetter;
+  final BoxDecoration? decoration;
+
+  VitComboBoxThemeData({
+    this.parentRenderBoxGetter,
+    this.decoration,
+  });
+}