+++ lib/vit_combo_box.dart
@@ -1,1 +1,3 @@
 export './components/combo_boxes/index.dart';
+export './theme/vit_combo_box_theme.dart';
+export './theme/vit_combo_box_theme_data.dart';